RavenDB-18277 - implementing /databases/*/debug/storage/report
